Consolidate fixtures
